% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UGridAndPoints}
\alias{UGridAndPoints}
\title{Represents a grid of tree loaded from XYZ file}
\description{
Most important method is CalculateFeaturesR
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructor \itemize{
\item FileNamePath: path to the XYZ file
\item swapYZ: TRUE/FALSE
\item cellSize: Size of the cell
}}

\item{\code{CalculateFeaturesR}}{Calculates features by using symmetry measure}

\item{\code{ReadFileToGridR}}{Reads the tree from XYZ file}

\item{\code{GetPointsR}}{Gets a list of all points of the XYZ file}
}}

